// lib/mongodb/database/model/event.model.ts
import { Document, Schema, model, models } from "mongoose";

// Define the Event interface
export interface IEvent extends Document {
  _id: string;
  title: string;
  description?: string;
  location?: string;
  createdAt: Date;
  imageUrl: string;
  startDateTime: Date;
  endDateTime: Date;
  price: string;
  isFree: boolean;
  url?: string;
  category: { _id: string, name: string };
  organizer: { _id: string, firstName: string, lastName: string };
}

// Define the Event Schema
const EventSchema = new Schema<IEvent>({
  title: { type: String, required: true },
  description: { type: String },
  location: { type: String },
  createdAt: { type: Date, default: Date.now },
  imageUrl: { type: String, required: true },
  startDateTime: { type: Date, default: Date.now },
  endDateTime: { type: Date, default: Date.now },
  price: { type: String },
  isFree: { type: Boolean, default: false },
  url: { type: String },
  category: { type: Schema.Types.ObjectId, ref: 'Category' },
  organizer: { type: Schema.Types.ObjectId, ref: 'User' }
});

// Register the Event model, checking if it already exists in Mongoose cache
const Event = models?.Event || model<IEvent>('Event', EventSchema);

export default Event;